requirements.txt
PyJWT==1.7.1
FLask-SQLAlchemy>= 2.5
requests==2.28.2
